created custom logic for negatory simple query builder serialization/deserialization
