+++ libs/designer-ui/src/lib/querybuilder/HybridQueryBuilder.tsx
@@ -1,7 +1,7 @@
 import { ValueSegmentType } from '../editor';
 import { Group } from './Group';
-import { GroupType } from './index';
-import type { GroupedItems, GroupItemProps, QueryBuilderProps, RowItemProps } from './index';
+import { checkHeights, getGroupedItems } from './helper';
+import type { GroupedItems, GroupItemProps, QueryBuilderProps } from './index';
 import { guid } from '@microsoft/utils-logic-apps';
 import { useFunctionalState, useUpdateEffect } from '@react-hookz/web';
 import { useEffect, useRef, useState } from 'react';
@@ -52,24 +52,4 @@ export const HybridQueryBuilderEditor = ({ getTokenPicker, groupProps, readonly,
       />
     </div>
   );
-};
-
-const checkHeights = (item: GroupItemProps | RowItemProps, returnVal: number[], height: number): number[] => {
-  if (item.checked) {
-    returnVal.push(height);
-  }
-  if (item.type === GroupType.GROUP) {
-    item.items.map((childItem) => checkHeights(childItem, returnVal, height + 1));
-  }
-  return returnVal;
-};
-
-const getGroupedItems = (item: GroupItemProps | RowItemProps, returnVal: GroupedItems[], index: number): GroupedItems[] => {
-  if (item.checked) {
-    returnVal.push({ item: { ...item, checked: false }, index: index });
-  }
-  if (item.type === GroupType.GROUP) {
-    item.items.map((childItem, index) => getGroupedItems(childItem, returnVal, index));
-  }
-  return returnVal;
 };
