+++ libs/designer-ui/src/lib/querybuilder/SimpleQueryBuilder.tsx
@@ -6,13 +6,15 @@ import type { ChangeHandler, ChangeState, GetTokenPickerHandler } from '../edito
 import { removeQuotes } from '../editor/base';
 import { StringEditor } from '../editor/string';
 import { Row } from './Row';
+import { getOperationValue, getOuterMostCommaIndex } from './helper';
 import type { IButtonStyles, IStyle } from '@fluentui/react';
 import { ActionButton, FontSizes } from '@fluentui/react';
-import { isNumber, isBoolean } from '@microsoft/parsers-logic-apps';
-import { guid } from '@microsoft/utils-logic-apps';
+import { guid, nthLastIndexOf } from '@microsoft/utils-logic-apps';
 import { useFunctionalState } from '@react-hookz/web';
 import { useEffect, useState } from 'react';
 import { useIntl } from 'react-intl';
+
+export * from './helper';
 
 export interface SimpleQueryBuilderProps {
   readonly?: boolean;
@@ -131,12 +133,21 @@ export const SimpleQueryBuilder = ({ getTokenPicker, itemValue, readonly, onChan
 };
 
 const convertRootPropToValue = (rootProps: RowItemProps): ValueSegment[] => {
-  const operandLiteral: ValueSegment = { id: guid(), type: ValueSegmentType.LITERAL, value: `@${rootProps.operator}(` };
-  const op1: ValueSegment = getOperationValue(rootProps.operand1[0]) ?? { id: guid(), type: ValueSegmentType.LITERAL, value: '' };
+  const { operator, operand1, operand2 } = rootProps;
+  const negatory = operator.includes('not');
+  const op1: ValueSegment = getOperationValue(operand1[0]) ?? { id: guid(), type: ValueSegmentType.LITERAL, value: '' };
   const separatorLiteral: ValueSegment = { id: guid(), type: ValueSegmentType.LITERAL, value: `,` };
-  const op2: ValueSegment = getOperationValue(rootProps.operand2[0]) ?? { id: guid(), type: ValueSegmentType.LITERAL, value: '' };
-  const endingLiteral: ValueSegment = { id: guid(), type: ValueSegmentType.LITERAL, value: `)` };
-  return [operandLiteral, op1, separatorLiteral, op2, endingLiteral];
+  const op2: ValueSegment = getOperationValue(operand2[0]) ?? { id: guid(), type: ValueSegmentType.LITERAL, value: '' };
+  if (negatory) {
+    const newOperator = operator.replace('not', '');
+    const negatoryOperatorLiteral: ValueSegment = { id: guid(), type: ValueSegmentType.LITERAL, value: `@not(${newOperator}(` };
+    const endingLiteral: ValueSegment = { id: guid(), type: ValueSegmentType.LITERAL, value: `))` };
+    return [negatoryOperatorLiteral, op1, separatorLiteral, op2, endingLiteral];
+  } else {
+    const operatorLiteral: ValueSegment = { id: guid(), type: ValueSegmentType.LITERAL, value: `@${operator}(` };
+    const endingLiteral: ValueSegment = { id: guid(), type: ValueSegmentType.LITERAL, value: `)` };
+    return [operatorLiteral, op1, separatorLiteral, op2, endingLiteral];
+  }
 };
 
 const convertAdvancedValueToRootProp = (value: ValueSegment[]): RowItemProps | undefined => {
@@ -146,23 +157,28 @@ const convertAdvancedValueToRootProp = (value: ValueSegment[]): RowItemProps | u
   const nodeMap = new Map<string, ValueSegment>();
   let stringValue = '';
   value.forEach((segment) => {
-    if (segment.type === ValueSegmentType.LITERAL) {
-      stringValue += segment.value;
-    } else if (segment.type === ValueSegmentType.TOKEN) {
-      stringValue += segment.value;
+    if (segment.type === ValueSegmentType.TOKEN) {
       nodeMap?.set(segment.value, segment);
     }
+    stringValue += segment.value;
   });
+  // cannot be converted into row format
   if (!stringValue.includes('@') || !stringValue.includes(',')) {
     return undefined;
   }
   let operator: string, operand1: ValueSegment[], operand2: ValueSegment[];
   try {
     operator = stringValue.substring(stringValue.indexOf('@') + 1, stringValue.indexOf('('));
+    const negatory = operator === 'not';
+    if (negatory) {
+      stringValue = stringValue.replace('@not(', '@');
+      operator = 'not' + stringValue.substring(stringValue.indexOf('@') + 1, stringValue.indexOf('('));
+    }
+    // if operator is not of the dropdownlist, it cannot be converted into row format
     if (!Object.values(RowDropdownOptions).includes(operator as RowDropdownOptions)) {
       return undefined;
     }
-    const operandSubstring = stringValue.substring(stringValue.indexOf('(') + 1, stringValue.lastIndexOf(')'));
+    const operandSubstring = stringValue.substring(stringValue.indexOf('(') + 1, nthLastIndexOf(stringValue, ')', negatory ? 2 : 1));
     const operand1String = removeQuotes(operandSubstring.substring(0, getOuterMostCommaIndex(operandSubstring)).trim());
     const operand2String = removeQuotes(operandSubstring.substring(getOuterMostCommaIndex(operandSubstring) + 1).trim());
     operand1 = [nodeMap.get(operand1String) ?? { id: guid(), type: ValueSegmentType.LITERAL, value: operand1String }];
@@ -170,43 +186,5 @@ const convertAdvancedValueToRootProp = (value: ValueSegment[]): RowItemProps | u
   } catch {
     return undefined;
   }
-
   return { operator, operand1, operand2, type: GroupType.ROW };
-};
-
-export const getOuterMostCommaIndex = (input: string): number => {
-  let outermostCommaIndex = -1;
-  let openParenthesesCount = 0;
-
-  for (let i = 0; i < input.length; i++) {
-    if (input[i] === '(') {
-      openParenthesesCount++;
-    } else if (input[i] === ')') {
-      openParenthesesCount--;
-    } else if (input[i] === ',' && openParenthesesCount === 0) {
-      outermostCommaIndex = i;
-      break;
-    }
-  }
-  return outermostCommaIndex;
-};
-
-const getOperationValue = (valSegment?: ValueSegment): ValueSegment | undefined => {
-  if (!valSegment) {
-    return undefined;
-  }
-  if (valSegment.token) {
-    return valSegment;
-  }
-  const currValue = valSegment.value;
-  const opeartionHasQuote = checkIfShouldHaveQuotes(valSegment);
-  return { id: valSegment.id, type: valSegment.type, value: `${opeartionHasQuote ? "'" : ''}${currValue}${opeartionHasQuote ? "'" : ''}` };
-};
-
-const checkIfShouldHaveQuotes = (valSegment: ValueSegment): boolean => {
-  const value = valSegment.value;
-  if (value && (isNumber(value) || isBoolean(value))) {
-    return false;
-  }
-  return true;
 };
